fixed faulty "Practice accumulating" problem set
Hi folks,

The only function the students have been taught that accumulates values in a list is `append()`, but it does not return a value, so the line:

> # List of word lengths: ["red", "green", "blue"] => [3, 5, 4]
> lengths = ____
> for word in ["red", "green", "blue"]:
>     lengths  = lengths.____(____)
> print(lengths)
> ~~~

will fail.  This can be done by using the `__add__()` function, but I'm pretty sure this isn't what the problem set designers had in mind.