Proposed changes to the CONTRIBUTING.md and design.md
In CONTRIBUTING.md there was a repetion of the word 'and'.

In design.md one of the questions required a verb to make sence. When challanges are presented the use of What instead of Which is counter intuitive in some cases. I understand that the questions are general although in some cases 'Which' seemes better. I only propose a change to which when is clear that one is refering to a choice among a smaller group of potentially known functions.