Changes proposed:
Make obvious that the menu to create a new notebook is the one that says "New".

In the episode of the types there is an unnecessary piece of mathmatics that will just make the exercise unclear. I belive the proposed solution is more intuitive to the audiance instead of more elaborate and technical mathmatics.